+++ doc/changelog.rst
@@ -1,6 +1,7 @@
 Changelog
 =========
 
+* :feature:`391` Add result.details, giving more options to adding/appending test details
 * :feature:`386` Make slash list support -f and other configuration parameters
 * :feature:`385` Add test details to xunit plugin output
 * :feature:`379` Allow exception marks to be used on both exception classes and exception values